Moved more views into util
